Ajout classes de noeuds et feuilles

+++ src/app5/ElemAST.java
@@ -9,7 +9,7 @@ public abstract class ElemAST {
   
   /** Evaluation d'AST
    */
-  public abstract int EvalAST();
+  public abstract double EvalAST();
 
 
   /** Lecture d'AST
